+++ Model-Modifier/src/scene/util/OrderVertices.cpp
@@ -7,10 +7,10 @@ float getAngle(glm::vec2 vert, glm::vec2 centroid)
 
 
 // order vertices by counter clockwise angle
-std::vector<unsigned int> OrderPolygonVertices(std::vector<glm::vec2> planePolygonVertices, glm::vec2 centroid, int numVerts)
+std::vector<unsigned int> OrderPolygonVertices(std::vector<glm::vec2> planePolygonVertices, glm::vec2 centroid, unsigned int numVerts)
 {
-	std::vector<std::pair<float, int>> angles_order(numVerts);
-	for (int i = 0; i < numVerts; i++)
+	std::vector<std::pair<float, unsigned int>> angles_order(numVerts);
+	for (unsigned int i = 0; i < numVerts; i++)
 	{
 		angles_order[i] = { getAngle(planePolygonVertices[i], centroid), i };
 	}
@@ -19,8 +19,8 @@ std::vector<unsigned int> OrderPolygonVertices(std::vector<glm::vec2> planePolyg
 	std::sort(angles_order.begin(), angles_order.end());
 
 	std::vector<unsigned int> vertexOrder(numVerts);
-	for (int j = 0; j < numVerts; j++) {
-		vertexOrder[j] = static_cast<unsigned int>(angles_order[j].second);
+	for (unsigned int j = 0; j < numVerts; j++) {
+		vertexOrder[j] = angles_order[j].second;
 	}
 	return vertexOrder;
 }
